+++ packages/ast/src/clients/rpc/class/grpc-gateway/grpc-gateway.ts
@@ -1,4 +1,4 @@
-import { GenericParseContext } from '../../../encoding';
+import { GenericParseContext } from '../../../../encoding';
 import { ProtoService, ProtoServiceMethod } from '@osmonauts/types';
 import * as t from '@babel/types'
