+++ packages/ast/src/clients/rpc/class/tendermint/rpc.arrowFunc.test.ts
@@ -1,16 +1,18 @@
-import { ProtoStore, traverse, getNestedProto } from '@osmonauts/proto-parser'
-import { defaultTelescopeOptions, ProtoService } from '@osmonauts/types';
-import { expectCode, getTestProtoStore, printCode } from '../../../../test-utils';
-import { GenericParseContext } from '../../../encoding';
+import { traverse, getNestedProto } from '@osmonauts/proto-parser'
+import { ProtoService } from '@osmonauts/types';
+import { expectCode, getTestProtoStore, printCode } from '../../../../../test-utils';
+import { GenericParseContext } from '../../../../encoding';
 import { createRpcInterface, createRpcClientClass, createRpcClientInterface } from './rpc';
-const store = getTestProtoStore();
+const store = getTestProtoStore({
+    classesUseArrowFunctions: true
+});
 store.traverseAll();
 
 it('RPC Query Client', () => {
     const ref = store.findProto('cosmos/auth/v1beta1/query.proto');
     const res = traverse(store, ref);
     const service: ProtoService = getNestedProto(res).Query;
-    const context = new GenericParseContext(ref, store, defaultTelescopeOptions);
+    const context = new GenericParseContext(ref, store, store.options);
     expectCode(createRpcClientInterface(context, service))
     expectCode(createRpcClientClass(context, service))
     expectCode(createRpcInterface(context, service))
