+++ packages/ast/src/clients/rpc/class/tendermint/rpc.msg.test.ts
@@ -1,7 +1,7 @@
 import { ProtoStore, traverse, getNestedProto } from '@osmonauts/proto-parser'
 import { defaultTelescopeOptions, ProtoService } from '@osmonauts/types';
-import { expectCode, getTestProtoStore, printCode } from '../../../../test-utils';
-import { GenericParseContext } from '../../../encoding';
+import { expectCode, getTestProtoStore, printCode } from '../../../../../test-utils';
+import { GenericParseContext } from '../../../../encoding';
 import { createRpcInterface, createRpcClientClass, createRpcClientInterface } from './rpc';
 const store = getTestProtoStore();
 store.traverseAll();
